Make MergeTree and Log table engines DROP and clean data

+++ src/Interpreters/DatabaseCatalog.cpp
@@ -121,11 +121,10 @@ TemporaryTableHolder::~TemporaryTableHolder()
 {
     if (id != UUIDHelpers::Nil)
     {
-        // auto table = getTable();
         /// Table maybe dropped before by DROP TEMPORARY TABLE
         /// Thus we need to verify the existence of the table before calling dropTable
         auto table = temporary_tables->tryGetTable("_tmp_" + toString(id), getContext());
-        if (!table)
+        if (table && !table->is_dropped)
         {
             table->flushAndShutdown();
             temporary_tables->dropTable(getContext(), "_tmp_" + toString(id));
